+++ app/src/main/java/com/vesaa/mytv/MyTVApplication.kt
@@ -8,6 +8,7 @@ import coil.Coil
 import coil.ImageLoader
 import coil.disk.DiskCache
 import coil.memory.MemoryCache
+import com.vesaa.mytv.BuildConfig
 import com.vesaa.mytv.data.work.EpgRefreshWorkScheduler
 import com.vesaa.mytv.ui.utils.SP
 
@@ -20,7 +21,9 @@ class MyTVApplication : Application() {
         AppGlobal.cacheDir = applicationContext.cacheDir
         SP.init(applicationContext)
 
-        initCoilImageLoader()
+        if (BuildConfig.CHANNEL_LOGOS_ENABLED) {
+            initCoilImageLoader()
+        }
 
         EpgRefreshWorkScheduler.schedule(this)
     }